small fixes

+++ GameProject/GameProject/ObjectContainer.cpp
@@ -8,7 +8,7 @@ void ObjectContainer::load(const sf::VideoMode &res)
 	const int max_objects = 4;
 	for (int i = 0; i < max_objects; ++i)
 	{
-		v.push_back(Object(res, 60, i * (float)(res.width / max_objects), color));
+		v.push_back(Object(res, 60, i * (float)(res.width) / (float)max_objects, color));
 	}
 }
